Widen Object::getMemSize() and overrides to size_t
Image::getMemSize() computed its size correctly but returned int, so
a true size beyond INT_MAX (reachable for a sufficiently large
in-memory image) silently truncated/wrapped to an arbitrary, possibly
negative value. Widen the whole getMemSize() chain (Object, Image,
Layer/LayerImage/LayerFolder, Sprite, Mask, Cel, CelData) from int to
size_t so undo-size accounting and the sprite properties dialog can't
observe a wrapped value.

Fixes #233

+++ src/doc/cel.h
@@ -58,7 +58,7 @@ public:
   void setPosition(const gfx::Point& pos);
   void setOpacity(int opacity);
 
-  virtual int getMemSize() const override
+  virtual size_t getMemSize() const override
   {
     return sizeof(Cel) + m_data->getMemSize();
   }
